update 2 script 7

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -10,7 +10,7 @@
 
 int print_last_digit(int n)
 {
-	int ret;
+	long int ret;
 
 	if (n < 0)
 	{
